requirements.txt
asgiref==3.7.2
attrs==23.1.0
Automat==22.10.0
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cmake==3.27.7
constantly==15.1.0
cryptography==41.0.3
cssselect==1.2.0
cx-Oracle==8.3.0
Django==4.2.4
django-import-export==3.3.1
djangorestframework==3.14.0
exceptiongroup==1.1.3
filelock==3.12.3
h11==0.14.0
huggingface-hub==0.17.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
kiwipiepy==0.16.0
kiwipiepy-model==0.16.0
lxml==4.9.3
nodeenv==1.8.0
numpy==1.23.1
outcome==1.2.0
packaging==23.1
parsel==1.8.1
Protego==0.3.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
PyMySQL==1.1.0
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.10.1
sentencepiece==0.1.99
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tldextract==3.4.4
torch==2.1.0
tqdm==4.65.0
transformers==4.34.0
trio==0.22.2
trio-websocket==0.10.3
Twisted==22.10.0
typing-extensions==4.7.1
urllib3==2.0.4
w3lib==2.1.2
wsproto==1.2.0
zope.interface==6.0
